Add support for binary file diffing
Added a new differ, `binary_diff`, which prints a single message if the
files differ in size or contents. Added a parameter to `new_goldenfile`
which allows the user to specify the differ to use. Updated tests and
example usage.

This is a breaking change. All calls to `new_goldenfile` must include a
new parameter, which should be `DifferType::Text` for regaining old
behavior.